+++ src/components/commitlist.rs
@@ -22,7 +22,6 @@ use crossterm::event::Event;
 use indexmap::IndexSet;
 use itertools::Itertools;
 use ratatui::{
-	backend::Backend,
 	layout::{Alignment, Rect},
 	style::Style,
 	text::{Line, Span},
@@ -534,7 +533,7 @@ impl CommitList {
 		let author = string_width_align(&e.author, author_width);
 
 		// commit author
-		txt.push(Span::styled::<String>(author, style_author));
+		txt.push(Span::styled(author, style_author));
 
 		txt.push(splitter.clone());
 
@@ -779,11 +778,7 @@ impl CommitList {
 }
 
 impl DrawableComponent for CommitList {
-	fn draw<B: Backend>(
-		&self,
-		f: &mut Frame<B>,
-		area: Rect,
-	) -> Result<()> {
+	fn draw(&self, f: &mut Frame, area: Rect) -> Result<()> {
 		let current_size = (
 			area.width.saturating_sub(2),
 			area.height.saturating_sub(2),