+++ src/components/compare_commits.rs
@@ -19,7 +19,6 @@ use asyncgit::{
 };
 use crossterm::event::Event;
 use ratatui::{
-	backend::Backend,
 	layout::{Constraint, Direction, Layout, Rect},
 	widgets::Clear,
 	Frame,
@@ -38,11 +37,7 @@ pub struct CompareCommitsComponent {
 }
 
 impl DrawableComponent for CompareCommitsComponent {
-	fn draw<B: Backend>(
-		&self,
-		f: &mut Frame<B>,
-		rect: Rect,
-	) -> Result<()> {
+	fn draw(&self, f: &mut Frame, rect: Rect) -> Result<()> {
 		if self.is_visible() {
 			let percentages = if self.diff.focused() {
 				(0, 100)