+++ src/components/fetch.rs
@@ -24,7 +24,6 @@ use asyncgit::{
 
 use crossterm::event::Event;
 use ratatui::{
-	backend::Backend,
 	layout::Rect,
 	text::Span,
 	widgets::{Block, BorderType, Borders, Clear, Gauge},
@@ -117,11 +116,7 @@ impl FetchComponent {
 }
 
 impl DrawableComponent for FetchComponent {
-	fn draw<B: Backend>(
-		&self,
-		f: &mut Frame<B>,
-		rect: Rect,
-	) -> Result<()> {
+	fn draw(&self, f: &mut Frame, rect: Rect) -> Result<()> {
 		if self.visible {
 			let progress = self.progress.unwrap_or_default().progress;