+++ src/components/file_revlog.rs
@@ -25,7 +25,6 @@ use chrono::{DateTime, Local};
 use crossbeam_channel::Sender;
 use crossterm::event::Event;
 use ratatui::{
-	backend::Backend,
 	layout::{Constraint, Direction, Layout, Rect},
 	text::{Line, Span, Text},
 	widgets::{Block, Borders, Cell, Clear, Row, Table, TableState},
@@ -378,7 +377,7 @@ impl FileRevlogComponent {
 		selection
 	}
 
-	fn draw_revlog<B: Backend>(&self, f: &mut Frame<B>, area: Rect) {
+	fn draw_revlog(&self, f: &mut Frame, area: Rect) {
 		let constraints = [
 			// type of change: (A)dded, (M)odified, (D)eleted
 			Constraint::Length(1),
@@ -391,8 +390,7 @@ impl FileRevlogComponent {
 		let title = self.get_title();
 		let rows = self.get_rows(now);
 
-		let table = Table::new(rows)
-			.widths(&constraints)
+		let table = Table::new(rows, constraints)
 			.column_spacing(1)
 			.highlight_style(self.theme.text(true, true))
 			.block(
@@ -465,11 +463,7 @@ impl FileRevlogComponent {
 }
 
 impl DrawableComponent for FileRevlogComponent {
-	fn draw<B: Backend>(
-		&self,
-		f: &mut Frame<B>,
-		area: Rect,
-	) -> Result<()> {
+	fn draw(&self, f: &mut Frame, area: Rect) -> Result<()> {
 		if self.visible {
 			let percentages = if self.diff.focused() {
 				(0, 100)