+++ src/components/stashmsg.rs
@@ -13,7 +13,7 @@ use crate::{
 use anyhow::Result;
 use asyncgit::sync::{self, RepoPathRef};
 use crossterm::event::Event;
-use ratatui::{backend::Backend, layout::Rect, Frame};
+use ratatui::{layout::Rect, Frame};
 
 pub struct StashMsgComponent {
 	repo: RepoPathRef,
@@ -24,11 +24,7 @@ pub struct StashMsgComponent {
 }
 
 impl DrawableComponent for StashMsgComponent {
-	fn draw<B: Backend>(
-		&self,
-		f: &mut Frame<B>,
-		rect: Rect,
-	) -> Result<()> {
+	fn draw(&self, f: &mut Frame, rect: Rect) -> Result<()> {
 		self.input.draw(f, rect)?;
 
 		Ok(())