+++ src/components/utils/scroll_vertical.rs
@@ -1,6 +1,6 @@
 use std::cell::Cell;
 
-use ratatui::{backend::Backend, layout::Rect, Frame};
+use ratatui::{layout::Rect, Frame};
 
 use crate::{
 	components::ScrollType,
@@ -109,12 +109,7 @@ impl VerticalScroll {
 		self.update(self.get_top(), line_count, visual_height)
 	}
 
-	pub fn draw<B: Backend>(
-		&self,
-		f: &mut Frame<B>,
-		r: Rect,
-		theme: &SharedTheme,
-	) {
+	pub fn draw(&self, f: &mut Frame, r: Rect, theme: &SharedTheme) {
 		draw_scrollbar(
 			f,
 			r,