+++ src/tabs/status.rs
@@ -83,9 +83,9 @@ pub struct Status {
 }
 
 impl DrawableComponent for Status {
-	fn draw<B: ratatui::backend::Backend>(
+	fn draw(
 		&self,
-		f: &mut ratatui::Frame<B>,
+		f: &mut ratatui::Frame,
 		rect: ratatui::layout::Rect,
 	) -> Result<()> {
 		let repo_unclean = self.repo_state_unclean();
@@ -200,9 +200,9 @@ impl Status {
 		}
 	}
 
-	fn draw_branch_state<B: ratatui::backend::Backend>(
+	fn draw_branch_state(
 		&self,
-		f: &mut ratatui::Frame<B>,
+		f: &mut ratatui::Frame,
 		chunks: &[ratatui::layout::Rect],
 	) {
 		if let Some(branch_name) = self.git_branch_name.last() {
@@ -281,9 +281,9 @@ impl Status {
 		}
 	}
 
-	fn draw_repo_state<B: ratatui::backend::Backend>(
+	fn draw_repo_state(
 		&self,
-		f: &mut ratatui::Frame<B>,
+		f: &mut ratatui::Frame,
 		r: ratatui::layout::Rect,
 	) {
 		if self.git_state != RepoState::Clean {